requirements.txt
pypdf2
bs4
requests
torch==2.3.1
pyyaml
transformers
